nbody: silence warning about implicit conversion from size_t to int

+++ nbody/nbody.cu
@@ -505,7 +505,7 @@ main( int argc, char *argv[] )
             }
         }
         for ( size_t i = 0; i < g_numGPUs; i++ ) {
-            gpuInit_struct initGPU = {i};
+            gpuInit_struct initGPU = {(int)i};
             g_GPUThreadPool[i].delegateSynchronous(
                 initializeGPU,
                 &initGPU );
